+++ src/main/java/com/diggydwarff/tobacconistmod/datagen/items/custom/pipeitems/CopperSmokingPipeItem.java
@@ -1,0 +1,12 @@
+package com.diggydwarff.tobacconistmod.datagen.items.custom.pipeitems;
+
+
+import com.diggydwarff.tobacconistmod.datagen.items.custom.WoodenSmokingPipeItem;
+
+public class CopperSmokingPipeItem extends WoodenSmokingPipeItem {
+
+    public CopperSmokingPipeItem(Properties properties) {
+        super(properties);
+    }
+
+}